+++ src/modules/users/dto/terms.dto.ts
@@ -3,7 +3,7 @@ import { ApiProperty } from '@nestjs/swagger';
 import { Type } from 'class-transformer';
 import { IsArray, IsBoolean, IsNumber, ValidateNested } from 'class-validator';
 
-export class TermsAgreeStateDto {
+export class SingleTermAgreeStateDto {
   @ApiProperty({
     description: '약관 ID (서버 → 클라이언트: JSON은 BigInt 그대로 직렬화 불가해서 string으로)',
   })
@@ -17,7 +17,7 @@ export class TermsAgreeStateDto {
 
 export class TermsHistoryItemDto {
   @ApiProperty({
-    description: '약관 ID (서버 → 클라이언트: JSON은 BigInt 그대로 직렬화 불가해서 string으로)',
+    description: '약관 ID',
   })
   termId!: string;
