+++ src/modules/users/services/oauth.users.service.ts
@@ -1,0 +1,4 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class OAuthUserService {}
