+++ src/modules/users/users.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 
+import { UsersService } from '@modules/users/services/users.service';
 import { UsersController } from '@modules/users/users.controller';
-import { UsersService } from '@modules/users/users.service';
 
 @Module({
   controllers: [UsersController],
